third_party/blink/web_tests/virtual/threaded-prefer-compositing/fast/scroll-behavior/smooth-scroll/main-thread-scrolling-reason-added-expected.txt

This test verifies that smooth scrolls initiated on the main thread add the appropriate main thread scrolling reason.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.getScrollAnimationState(document) is RUNNING_ON_COMPOSITOR
PASS internals.mainThreadScrollingReasons(document) is ANIMATING_TEXT
PASS internals.getScrollAnimationState(document) became IDLE
PASS document.scrollingElement.scrollTop is 40
PASS internals.mainThreadScrollingReasons(document) == '' is true
PASS successfullyParsed is true

TEST COMPLETE
